updated frank's data bnn

+++ final_repo_test/final_repo_test/Models/CaseTable.cs
@@ -7,7 +7,7 @@ namespace final_repo_test.Models
         [Key]
         public int Case_ID { get; set; } = default!;
         public string Case_Name { get; set; } = default!;
-        public decimal Case_PricePerDay { get; set; } = default!;
+        public int Case_PricePerDay { get; set; } = default!;
         public ICollection<Ads>? Ads { get; set; }
         public ICollection<Map_ADArea>? Map_ADAreas { get; set; }
